add serialization class
