+++ TaskMaster/CMakeLists.txt
@@ -14,9 +14,10 @@ add_library(TaskMaster src/Project/ProjectController.cpp
                         src/models/Project.cpp
                         src/DbDriver/DbDriver.cpp
                         src/MainPadge/MainPadgeController.cpp
+                        src/Serialization/Serialization.cpp
                         )
 
-target_include_directories(TaskMaster PUBLIC include/Project include/Board include/Task include/User include/models include/DbDriver include/MainPadge)
+target_include_directories(TaskMaster PUBLIC include/Project include/Board include/Task include/User include/models include/DbDriver include/MainPadge include/Serialization)
 
 if(BUILD_TESTS)
   add_subdirectory(tests)
